// Frontend/ShoppingCart/src/app/models/User/IUserModel.ts
/**
 * Represents the structure for a unique user.
 */
export interface IUserModel {
  /// <summary>
  /// Autogenerated ID that uniquely identifies each user with their unique ID.
  /// </summary>
  /// <example>1</example>
  UserID: number;
  /// <summary>
  /// Username of the user.
  /// </summary>
  /// <example>karthik</example>
  Username: string;
  /// <summary>
  /// Password of the user.
  /// </summary>
  /// <example>p@ssw0rd123!</example>
  Password: string;
  /// <summary>
  /// Email address of the user.
  /// </summary>
  /// <example>karthik@gmail.com</example>
  Email: string;
  /// <summary>
  /// Phone numer of the user.
  /// </summary>
  /// <example>9940568874</example>
  Phone: number;
}